+++ aniyomi-domain/src/main/java/naiko/domain/extensionrepo/anime/interactor/DeleteAnimeExtensionRepo.kt
@@ -1,0 +1,11 @@
+package naiko.domain.extensionrepo.anime.interactor
+
+import naiko.domain.extensionrepo.anime.repository.AnimeExtensionRepoRepository
+
+class DeleteAnimeExtensionRepo(
+    private val repository: AnimeExtensionRepoRepository,
+) {
+    suspend fun await(baseUrl: String) {
+        repository.deleteRepo(baseUrl)
+    }
+}