+++ aniyomi-domain/src/main/java/naiko/domain/extensionrepo/manga/interactor/DeleteMangaExtensionRepo.kt
@@ -1,0 +1,11 @@
+package naiko.domain.extensionrepo.manga.interactor
+
+import naiko.domain.extensionrepo.manga.repository.MangaExtensionRepoRepository
+
+class DeleteMangaExtensionRepo(
+    private val repository: MangaExtensionRepoRepository,
+) {
+    suspend fun await(baseUrl: String) {
+        repository.deleteRepo(baseUrl)
+    }
+}