+++ aniyomi-domain/src/main/java/naiko/domain/extensionrepo/manga/interactor/GetMangaExtensionRepoCount.kt
@@ -1,0 +1,9 @@
+package naiko.domain.extensionrepo.manga.interactor
+
+import naiko.domain.extensionrepo.manga.repository.MangaExtensionRepoRepository
+
+class GetMangaExtensionRepoCount(
+    private val repository: MangaExtensionRepoRepository,
+) {
+    fun subscribe() = repository.getCount()
+}