+++ app/src/main/java/eu/kanade/tachiyomi/source/LocalSource.kt
@@ -47,7 +47,7 @@ import yokai.util.lang.getString
 class LocalSource(private val context: Context) : CatalogueSource, UnmeteredSource {
     companion object {
         const val ID = 0L
-        const val HELP_URL = "https://mihon.app/docs/guides/local-source/"
+        const val HELP_URL = "https://naiko.app/docs/guides/local-source/"
 
         private const val COVER_NAME = "cover.jpg"
         private val LATEST_THRESHOLD = TimeUnit.MILLISECONDS.convert(7, TimeUnit.DAYS)