+++ app/src/main/java/yokai/presentation/settings/screen/SettingsDataScreen.kt
@@ -333,4 +333,4 @@ object SettingsDataScreen : ComposableSettings {
     }
 }
 
-const val BACKUPS_HELP_URL = "https://mihon.app/docs/guides/backups"
+const val BACKUPS_HELP_URL = "https://naiko.app/docs/guides/backups"