+++ app/src/main/java/yokai/presentation/webview/WebViewScreenContent.kt
@@ -202,7 +202,7 @@ fun WebViewScreenContent(
                                     .clip(MaterialTheme.shapes.small)
                                     .clickable {
                                         uriHandler.openUri(
-                                            "https://mihon.app/docs/guides/troubleshooting/#cloudflare",
+                                            "https://naiko.app/docs/guides/troubleshooting/#cloudflare",
                                         )
                                     },
                             )